working forreal this time

+++ Population.cc
@@ -11,7 +11,7 @@
 Population :: Population()
 {
 	popSize = 10;
-	int fittest = 0;
+	double fittest = 0.0;
 }
 
 //Initialize population
@@ -31,8 +31,8 @@ void Population :: initializePopulation(int _popSize, int _geneSize, int _atomCo
 //Get the fittest individual
 int Population :: getFittest() 
 {
-	int maxFit = INT_MIN;
-	int maxFitIndex = 0;
+	double maxFit = INT_MIN;
+	double maxFitIndex = 0;
 	for (int i = 0; i < popSize; i++) 
 	{
 		if (maxFit <= individuals[i].fitness) 
@@ -48,8 +48,8 @@ int Population :: getFittest()
 //Get the second most fittest individual
 int Population :: getSecondFittest() 
 {
-	int maxFit1 = 0;
-	int maxFit2 = 0;
+	double maxFit1 = 0;
+	double maxFit2 = 0;
 	for (int i = 0; i < popSize; i++) 
 	{
 		if (individuals[i].fitness > individuals[maxFit1].fitness) 
@@ -68,8 +68,8 @@ int Population :: getSecondFittest()
 //Get index of least fittest individual
 int Population :: getLeastFittestIndex() 
 {
-	int minFitVal = INT_MAX;
-	int minFitIndex = 0;
+	double minFitVal = INT_MAX;
+	double minFitIndex = 0;
 	for (int i = 0; i < popSize; i++) 
 	{
 		if (minFitVal >= individuals[i].fitness)
